Backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
firebase-admin==6.1.0
requests==2.28.2
pika==1.3.1
cryptography==39.0.2
qrcode==7.4.2
sendgrid==6.10.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.10.0